requirements.txt
# Core dependencies
discord.py>=2.3.0
aiohttp>=3.9.0
pyyaml>=6.0
aiosqlite>=0.19.0

# Voice dependencies (CPU-only for AMD)
SpeechRecognition>=3.10.0
edge-tts>=6.1.0
pyttsx3>=2.90
pyaudio>=0.2.14
pocketsphinx>=5.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For OpenAI provider support
# openai>=1.0.0